requirements.txt
fastapi==0.96.0
uvicorn==0.22.0
pymongo==4.4.0
passlib==1.7.4
python-dotenv==1.0.0


